@@ -1,27 +1,30 @@
-const { getConfig, getLogger } = require('boiler');
-const { bridgeConnection, streamIdForUserId, getUserParentStreamId, logErrorOnBridgeAccount } = require('../lib/bridgeAccount');
-const pryvService = require('../lib/pryvService');
-const { internalError, badRequest, serviceError } = require('../errors');
-const user = require('./user.js');
+import boiler from 'boiler';
+import { bridgeConnection, streamIdForUserId, getUserParentStreamId, logErrorOnBridgeAccount } from '../lib/bridgeAccount.ts';
+import * as pryvService from '../lib/pryvService.ts';
+import { internalError, badRequest, serviceError } from '../errors/index.ts';
+import type { AppError } from '../errors/index.ts';
+import * as user from './user.ts';
+import ShortUniqueId from 'short-unique-id';
+import { advertiseNewUserToPlugins, requiredPermissionsAndStreams } from '../lib/plugins.ts';
 
-const ShortUniqueId = require('short-unique-id');
-const { advertiseNewUserToPlugins, requiredPermissionsAndStreams } = require('../lib/plugins.js');
+const { getConfig, getLogger } = boiler;
+let _logger: ReturnType<typeof getLogger> | null = null;
+function logger () { return _logger || (_logger = getLogger('onboard')); }
+
 const onboardingSecretGenerator = new ShortUniqueId({ dictionary: 'alphanum', length: 24 });
-
-const logger = getLogger('onboard');
-
-module.exports = {
-  init,
-  initiate,
-  finalize,
-  authStatusesGet,
-  authStatusesClean
-};
 
 /**
  * Will be set by init with values from the config and service
  */
-const settings = {
+const settings: {
+  requestingAppId: string | null;
+  requestedPermissions: unknown[] | null;
+  apiAccessURL: string | null;
+  returnURL: string | null;
+  consentMessage: string | null;
+  partnerURLs: any;
+  ensureBaseStreams?: unknown;
+} = {
   requestingAppId: null,
   requestedPermissions: null,
   apiAccessURL: null,
@@ -29,30 +32,41 @@ const settings = {
   consentMessage: null,
   partnerURLs: null
 };
+
 // from Pryv service
-async function init () {
+async function init (): Promise<void> {
   const config = await getConfig();
-  settings.requestingAppId = config.get('service:appId');
-  settings.consentMessage = config.get('service:consentMessage');
-  settings.returnURL = config.get('baseURL') + '/user/onboard/finalize/';
+  settings.requestingAppId = config.get<string>('service:appId');
+  settings.consentMessage = config.get<string>('service:consentMessage');
+  settings.returnURL = config.get<string>('baseURL') + '/user/onboard/finalize/';
 
   settings.apiAccessURL = (await pryvService.service().info()).access;
   settings.partnerURLs = config.get('partnerURLs');
 
   // add permissions and streams from plugins
-  const permissionsFromSettings = config.get('service:userPermissionRequest');
+  const permissionsFromSettings = config.get<unknown[]>('service:userPermissionRequest');
   validatePermissions(permissionsFromSettings);
   const { permissions, streams } = requiredPermissionsAndStreams(permissionsFromSettings);
-  settings.requestedPermissions = permissions;
+  settings.requestedPermissions = permissions as unknown[];
   settings.ensureBaseStreams = streams;
+}
+
+interface RedirectURLs {
+  success: string;
+  cancel: string;
+}
+
+interface InitiateResult {
+  redirectURLs: RedirectURLs;
+  webhookClientData: Record<string, unknown>;
+  responseBody: any;
+  onboardingSecret: string;
 }
 
 /**
  * Create an onboarding URL for this patient
- * @param {string} partnerUserId
- * @returns {string} URL to onboard the patient
  */
-async function initiate (partnerUserId, redirectURLs, webhookClientData) {
+async function initiate (partnerUserId: string, redirectURLs: RedirectURLs, webhookClientData: Record<string, unknown>): Promise<unknown> {
   // check if user is active
 
   const userStatus = await user.status(partnerUserId, false);
@@ -79,7 +93,7 @@ async function initiate (partnerUserId, redirectURLs, webhookClientData) {
             }
       }
   };
-  const response = await fetch(settings.apiAccessURL, {
+  const response = await fetch(settings.apiAccessURL!, {
     method: 'POST',
     headers: {
       Accept: 'application/json',
@@ -92,13 +106,13 @@ async function initiate (partnerUserId, redirectURLs, webhookClientData) {
   const onboardingSecret = onboardingSecretGenerator.randomUUID();
 
   // -- store request intent
-  const initiateResult = { redirectURLs, webhookClientData, responseBody, onboardingSecret };
+  const initiateResult: InitiateResult = { redirectURLs, webhookClientData, responseBody, onboardingSecret };
   await authStatusStore(partnerUserId, initiateResult);
 
   const result = {
     type: 'authRequest',
     onboardingSecret,
-    redirectUserURL: responseBody.url,
+    redirectUserURL: (responseBody as any).url,
     context: initiateResult
   };
 
@@ -107,16 +121,13 @@ async function initiate (partnerUserId, redirectURLs, webhookClientData) {
 
 /**
  * Finalize an onboarding process
- * @param {string} partnerUserId
- * @param {string} pollParam
- * @returns url
  */
-async function finalize (partnerUserId, pollParam) {
+async function finalize (partnerUserId: string, pollParam: string | string[]): Promise<string> {
   try {
     return await finalizeToBeCatched(partnerUserId, pollParam);
-  } catch (e) {
+  } catch (e: any) {
     if (!e.skipWebHookCall) {
-      const webhookParams = Object.assign(
+      const webhookParams: Record<string, unknown> = Object.assign(
         {
           type: 'ERROR',
           partnerUserId,
@@ -125,7 +136,7 @@ async function finalize (partnerUserId, pollParam) {
         }, e.webhookParams || { });
       // call webhook
       await webhookCall(settings.partnerURLs.webhookOnboard, webhookParams);
-      logger.error(e);
+      logger().error(e);
     }
     const errorObject = {
       partnerUserId,
@@ -140,26 +151,23 @@ async function finalize (partnerUserId, pollParam) {
 
 /**
  * really finalized
- * @param {string} partnerUserId
- * @param {string} pollParam
- * @returns url - to redirect the user
  */
-async function finalizeToBeCatched (partnerUserId, pollParam) {
+async function finalizeToBeCatched (partnerUserId: string, pollParam: string | string[]): Promise<string> {
   // might be an Array ..
   const pollURL = Array.isArray(pollParam) ? pollParam[0] : pollParam;
   if (pollURL == null || !pollURL.startsWith('http')) badRequest('Missing or invalid prYvpoll URL');
-  const pollContent = await (await fetch(pollURL)).json();
+  const pollContent = await (await fetch(pollURL)).json() as any;
 
   // safety check that onboard process has started
   const currentAuthStatuses = await authStatusesGet(partnerUserId);
-  const matchingStatuses = currentAuthStatuses.filter(s => s.content.responseBody.poll === pollURL);
+  const matchingStatuses = currentAuthStatuses.filter((s: any) => s.content.responseBody.poll === pollURL);
   if (matchingStatuses.length !== 1) {
-    logger.error('No matching pending request for this user', { partnerUserId, pollParam });
+    logger().error('No matching pending request for this user', { partnerUserId, pollParam });
     // -- redirect to partner error page
     return getErrorRedirectURLWithMessage('No matching pending request');
   }
   const matchingStatusContent = matchingStatuses[0].content;
-  const webhookParams = Object.assign({ partnerUserId, onboardingSecret: matchingStatusContent.onboardingSecret }, matchingStatusContent.webhookClientData);
+  const webhookParams: Record<string, unknown> = Object.assign({ partnerUserId, onboardingSecret: matchingStatusContent.onboardingSecret }, matchingStatusContent.webhookClientData);
 
   // REMOVE pending request in background
   process.nextTick(() => { authStatusesClean(currentAuthStatuses); });
@@ -181,7 +189,7 @@ async function finalizeToBeCatched (partnerUserId, pollParam) {
     // CANCELLED
     webhookParams.type = 'CANCEL';
     webhookParams.status = pollContent.status;
-  } catch (e) {
+  } catch (e: any) {
     // just attach webhookParams to forward the error
     e.webhookParams = webhookParams;
     throw e;
@@ -193,26 +201,20 @@ async function finalizeToBeCatched (partnerUserId, pollParam) {
 // ------ onboard steps
 
 /**
- * Set the
- */
-
-/**
  * Get pending auth status (may be sevrals)
- * @param {string} partnerUserId
- * @returns {Array} of status
  */
-async function authStatusesGet (partnerUserId) {
+async function authStatusesGet (partnerUserId: string): Promise<any[]> {
   const userStreamId = streamIdForUserId(partnerUserId);
   const apiCalls = [{
     method: 'events.get',
     params: { streams: [userStreamId], types: ['temp-status/bridge-auth-request'] }
   }];
-  const response = (await bridgeConnection().api(apiCalls))[0];
+  const response = (await bridgeConnection().api(apiCalls as any) as any)[0];
   // -- todo check response
   return response.events || [];
 }
 
-async function authStatusStore (partnerUserId, payload) {
+async function authStatusStore (partnerUserId: string, payload: unknown): Promise<unknown> {
   const userStreamId = streamIdForUserId(partnerUserId);
   const apiCalls = [{
     method: 'streams.create',
@@ -225,26 +227,25 @@ async function authStatusStore (partnerUserId, payload) {
       content: payload
     }
   }];
-  const response = await bridgeConnection().api(apiCalls);
+  const response: any = await bridgeConnection().api(apiCalls as any);
   if (!response[1].event || response[1].error) serviceError('Failed storing auth status', response[1]);
   return response[1].event;
 }
 
 /**
  * Array of pending authStatus to remove
- * @param {Array<Events>} authStatusEvents
  */
-async function authStatusesClean (authStatusEvents) {
+async function authStatusesClean (authStatusEvents: any[]): Promise<void> {
   if (!authStatusEvents || authStatusEvents.length < 1) return;
-  const apiCalls = [];
+  const apiCalls: any[] = [];
   for (const e of authStatusEvents) {
     const deleteCall = { method: 'events.delete', params: { id: e.id } };
     apiCalls.push(deleteCall, deleteCall); // twice for a real delete
   }
-  const res = await bridgeConnection().api(apiCalls);
+  const res: any = await bridgeConnection().api(apiCalls as any);
   for (const r of res) {
-    if (r.error) logger.error('Failed deleting status event', r);
-    if (r.eventDeletion) logger.info(`Deleted status event id: ${r.eventDeletion.id}`);
+    if (r.error) logger().error('Failed deleting status event', r);
+    if (r.eventDeletion) logger().info(`Deleted status event id: ${r.eventDeletion.id}`);
   }
 }
 
@@ -253,7 +254,7 @@ async function authStatusesClean (authStatusEvents) {
 /**
  * Validate if settings for requested permissions is valid
  */
-function validatePermissions (permissions) {
+function validatePermissions (permissions: any): void {
   if (!Array.isArray(permissions)) internalError('Permissions setting should be an array: ' + JSON.stringify(permissions, null, 2));
   if (permissions.length === 0) internalError('Permissions setting should have one element ' + JSON.stringify(permissions, null, 2));
   for (const p of permissions) {
@@ -265,24 +266,24 @@ function validatePermissions (permissions) {
 }
 
 // --- webhook caller
-async function webhookCall (whSettings, params) {
+async function webhookCall (whSettings: any, params: Record<string, unknown>): Promise<void> {
   // Call partner webhook
-  const fetchParams = {
+  const fetchParams: RequestInit = {
     method: whSettings.method,
-    headers: whSettings.headers
+    headers: { ...whSettings.headers }
   };
   let queryParams = '';
   if (whSettings.method === 'GET') {
-    if (Object.keys(params).length > 0) queryParams = '?' + new URLSearchParams(params);
+    if (Object.keys(params).length > 0) queryParams = '?' + new URLSearchParams(params as Record<string, string>);
   } else { // assume POST
-    fetchParams.headers['Content-Type'] = 'application/json';
+    (fetchParams.headers as Record<string, string>)['Content-Type'] = 'application/json';
     fetchParams.body = JSON.stringify(params);
   }
   try {
     await fetch(whSettings.url + queryParams, fetchParams);
   } catch (e) {
-    logger.error('Failed contacting partner backend', e);
-    const e2 = new Error('Failed contacting partner backend');
+    logger().error('Failed contacting partner backend', e);
+    const e2: AppError = new Error('Failed contacting partner backend');
     e2.errorObject = {
       webhookCall: {
         whSettings,
@@ -295,7 +296,9 @@ async function webhookCall (whSettings, params) {
   }
 }
 
-function getErrorRedirectURLWithMessage (message) {
+function getErrorRedirectURLWithMessage (message: string): string {
   const encodedMessage = encodeURIComponent(message);
   return settings.partnerURLs.defaultRedirectOnError + '?message=' + encodedMessage;
 }
+
+export { init, initiate, finalize, authStatusesGet, authStatusesClean };
