@@ -1,23 +1,33 @@
-const { bridgeConnection, streamIdForUserId, getUserParentStreamId, getActiveUserStreamId } = require('../lib/bridgeAccount');
-const { unkownRessource, serviceError, badRequest } = require('../errors');
-const { pryv } = require('hds-lib');
+import { bridgeConnection, streamIdForUserId, getUserParentStreamId, getActiveUserStreamId } from '../lib/bridgeAccount.ts';
+import { unkownRessource, serviceError, badRequest } from '../errors/index.ts';
+import { pryv } from 'hds-lib';
 
-module.exports = {
-  status,
-  exists,
-  addCredentialToBridgeAccount,
-  getPryvConnectionAndStatus,
-  setStatus,
-  getAllUsersApiEndpoints
-};
+interface UserInfo {
+  active: boolean;
+  partnerUserId: string;
+  apiEndpoint: string;
+  created: number;
+  modified: number;
+}
+
+interface SyncStatusInfo {
+  content: unknown;
+  lastSync: number | undefined;
+}
+
+interface UserStatus {
+  user: UserInfo;
+  syncStatus: SyncStatusInfo;
+}
+
+interface StatusAndPryvConnection extends UserStatus {
+  connection: InstanceType<typeof pryv.Connection>;
+}
 
 /**
  * Add user credentials to partner account
- * @param {string} partnerUserId
- * @param {string} appApiEndpoint
- * @returns {Event} - the created event content
  */
-async function addCredentialToBridgeAccount (partnerUserId, appApiEndpoint) {
+async function addCredentialToBridgeAccount (partnerUserId: string, appApiEndpoint: string): Promise<unknown> {
   const streamUserId = streamIdForUserId(partnerUserId);
   const apiCalls = [{
     method: 'streams.create',
@@ -26,43 +36,26 @@ async function addCredentialToBridgeAccount (partnerUserId, appApiEndpoint) {
     method: 'events.create',
     params: { streamIds: [streamUserId, getActiveUserStreamId()], type: 'credentials/pryv-api-endpoint', content: appApiEndpoint }
   }];
-  const result = await bridgeConnection().api(apiCalls);
+  const result: any = await bridgeConnection().api(apiCalls as any);
   if (result[1]?.error?.id) throw serviceError('Failed add user credentials', result[1]);
   return result[1];
 }
 
-async function exists (partnerUserId) {
+async function exists (partnerUserId: string): Promise<boolean> {
   const streamUserId = streamIdForUserId(partnerUserId);
   const apiCalls = [{
     method: 'events.get',
     params: { streams: [streamUserId], limit: 1, types: ['credentials/pryv-api-endpoint'] }
   }];
-  const result = await bridgeConnection().api(apiCalls);
+  const result: any = await bridgeConnection().api(apiCalls as any);
   if (result[0]?.error?.id === 'unknown-referenced-resource') return false;
   return true;
 }
 
 /**
- * @typedef {UserStatus}
- * @property {Object} user
- * @property {boolean} user.active
- * @property {string} user.partnerUserId
- * @property {string} user.apiEndpoint
- * @property {number} user.created - EPOCH time in seconds
- * @property {number} user.modified - EPOCH time in seconds
- * @property {Object} [syncStatus]
- * @property {Object} syncStatus.content
- * @property {number} syncStatus.lastSync - EPOCH time in seconds
- */
-
-/**
  * Get user status
- * @param {string} partnerUserId
- * @param {boolean} [throwUnkown=true] - if truethrow an error, otherwise return null
- * @returns {UserStatus|null}
- * @throws 404 Unkown User
  */
-async function status (partnerUserId, throwUnkown = true) {
+async function status (partnerUserId: string, throwUnkown = true): Promise<UserStatus | null> {
   const streamUserId = streamIdForUserId(partnerUserId);
   const apiCalls = [{
     method: 'events.get',
@@ -71,7 +64,7 @@ async function status (partnerUserId, throwUnkown = true) {
     method: 'events.get',
     params: { streams: [streamUserId], limit: 1, types: ['sync-status/bridge'] }
   }];
-  const resultFromBC = await bridgeConnection().api(apiCalls);
+  const resultFromBC: any = await bridgeConnection().api(apiCalls as any);
   if (resultFromBC[0]?.error?.id === 'unknown-referenced-resource') {
     if (throwUnkown) {
       unkownRessource('Unkown user', { userId: partnerUserId });
@@ -88,7 +81,7 @@ async function status (partnerUserId, throwUnkown = true) {
     }
     return null;
   }
-  const result = {
+  const result: UserStatus = {
     user: {
       active: userEvent.streamIds.includes(getActiveUserStreamId()),
       partnerUserId,
@@ -104,19 +97,19 @@ async function status (partnerUserId, throwUnkown = true) {
   return result;
 }
 
-async function setStatus (partnerUserId, active) {
+async function setStatus (partnerUserId: string, active?: boolean): Promise<{ active: boolean }> {
   const streamUserId = streamIdForUserId(partnerUserId);
   const apiCalls = [{
     method: 'events.get',
     params: { streams: [streamUserId], limit: 1, types: ['credentials/pryv-api-endpoint'] }
   }];
-  const resultFromBC = await bridgeConnection().api(apiCalls);
+  const resultFromBC: any = await bridgeConnection().api(apiCalls as any);
   if (resultFromBC[0]?.error?.id === 'unknown-referenced-resource') unkownRessource('Unkown user', { userId: partnerUserId });
   const error = resultFromBC.error || resultFromBC[1]?.error;
   if (error) serviceError('Failed to get user status', error);
   const userEvent = resultFromBC[0].events[0];
   const currentStatus = userEvent.streamIds.includes(getActiveUserStreamId());
-  if (currentStatus === active) return { active };
+  if (currentStatus === active) return { active: currentStatus };
 
   // change streams
   const newStreamIds = [...userEvent.streamIds];
@@ -135,7 +128,7 @@ async function setStatus (partnerUserId, active) {
       }
     }
   }];
-  const resultUpdate = await bridgeConnection().api(apiCallsUpdate);
+  const resultUpdate: any = await bridgeConnection().api(apiCallsUpdate as any);
   if (resultUpdate[0]?.error?.id === 'unknown-referenced-resource') unkownRessource('Unkown user', { userId: partnerUserId });
   const errorUpdate = resultUpdate.error || resultUpdate[0]?.error;
   if (errorUpdate) serviceError('Failed to get user status', errorUpdate);
@@ -144,37 +137,34 @@ async function setStatus (partnerUserId, active) {
 }
 
 /**
- * @typedef {Object} StatusAndPryvConnection
- * @augments UserStatus
- * @property {connection} connection
- */
-
-/**
  * Pryv API endpoint and status for the user
- * @param {string} partnerUserId
- * @param {boolean} [includesInactive=false] - if true, include inactive users
- * @returns {StatusAndPryvConnection}
- * @throws 404 Unkown User
- * @throws 400 Deactivated User - if includesInactive is false & user is deactivated
  */
-async function getPryvConnectionAndStatus (partnerUserId, includesInactive = false) {
+async function getPryvConnectionAndStatus (partnerUserId: string, includesInactive = false): Promise<StatusAndPryvConnection> {
   const statusResult = await status(partnerUserId);
-  if (!statusResult.user.active && !includesInactive) {
+  if (!statusResult!.user.active && !includesInactive) {
     badRequest('Deactivated User', { userId: partnerUserId });
   }
-  const connection = new pryv.Connection(statusResult.user.apiEndpoint);
+  const connection = new pryv.Connection(statusResult!.user.apiEndpoint);
   return {
-    ...statusResult,
+    ...statusResult!,
     connection
   };
 }
 
 /**
  * Get all users and their status
- * @param {callback} forEachEvent - called for Each event
  */
-async function getAllUsersApiEndpoints (forEachEvent) {
+async function getAllUsersApiEndpoints (forEachEvent: (event: unknown) => void): Promise<unknown> {
   const now = (new Date()).getTime() / 1000;
   const queryParams = { fromTime: 0, toTime: now, streams: [getUserParentStreamId()], types: ['credentials/pryv-api-endpoint'] };
   return await bridgeConnection().getEventsStreamed(queryParams, forEachEvent);
 }
+
+export {
+  status,
+  exists,
+  addCredentialToBridgeAccount,
+  getPryvConnectionAndStatus,
+  setStatus,
+  getAllUsersApiEndpoints
+};
